Rename component to avoid clashes with reserved keywords
