+++ src/components/location-map/index.jsx
@@ -1,0 +1,1 @@
+export { default } from "./location-map";
