+++ src/components/location-map/map-loading-state.stories.jsx
@@ -1,0 +1,9 @@
+import MapLoadingState from "./map-loading-state";
+
+export default {
+  title: "Components/LocationMap/MapLoadingState",
+  component: MapLoadingState,
+  tags: ["autodocs"],
+};
+
+export const Default = {};
