+++ src/components/location-map/map-pin.jsx
